FileUtils, Engine refactor
