+++ src/main/java/org/j2overhead/m1ke/M1keRunner.java
@@ -1,7 +1,10 @@
 package org.j2overhead.m1ke;
 
-public class Main {
+public class M1keRunner {
     public static void main(String[] args) {
 
+    }
+
+    void init(String[] args) {
     }
 }
