--- src/main/java/org/j2overhead/m1ke/model/Origin.java
@@ -1,4 +1,0 @@
-package org.j2overhead.m1ke.model;
-
-public class Origin {
-}
