+++ src/main/java/org/j2overhead/m1ke/model/Repository.java
@@ -1,0 +1,7 @@
+package org.j2overhead.m1ke.model;
+
+import java.util.ArrayList;
+
+public class Repository {
+    ArrayList<Branch> branches;
+}
